Correct Uemis extra data
When dropping the pointless surface samples at the end of the dive we
stopped going through all the samples - for that reason the offset at the
end to find the data in the footer was wrong.

Now the extra data all seems to match what I can see on my screen after
the dive.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>