Don't skip the first tag when filtering
No idea why this code was there... but what it caused was an invalid
access to checkState[-1] when a dive had that first tag. This explains why
sometimes the dives that had that tag were shown, regardless of whether
the selected filter criteria should have excluded them or not.

Fixes #769

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>