Modified the UI of the User Survey Page
The UI in the User Survey page was not clear and the Text Blocks were getting
cut. So I increased the size of the Text Blocks.

Also the survey was using CheckBoxes instead of RadioButtons. The problem with
this is that the user could choose to be both Recreational and Technical diver
which cannot be possible. So I replaced the CheckBoxes with RadioButtons.

I also replaced the CheckBoxes, which asked the user from where he was importing data, with RadioButtons so that the user can select only one option.

I also modified the English used and made it concise and clear.

Signed-off by : Nikhil Bharadwaj Gosala <nikhil.gosala@gmail.com>

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>