Add the current OBS spec file for daily builds
This is simply here for people to look at. It will age immediately and it makes
no sense to try to keep it current here as it is maintained in OBS. But I think
it might be a useful starting point for others who want to package daily builds
of Subsurface.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>