User-manual: Dive planner text
I took Robert's contribution describing the dive planner and expanded it a
bit, added two more graphics to show CCR and pSCR dive planning, and
changed a third image.

Moved the section dealing with configuring a dive computer to be in a more
logical place, just before the Subsurface Settings.

Signed-off-by: willem ferguson <willemferguson@zoology.up.ac.za>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>